Use <C-f> for Emmet instead of <C-y>
I always like to learn defaults instead of re-mapping things, so this is
a somewhat hard decision for me :P
Emmet it pretty cool, but it needs a key mapping that can be used in normal,
insert and visual mode.

<leader> does not work in insert mode, so I'd have to leave insert mode
to use it. This breaks the workflow a little when using many commands in
a quick succession.

<C-y> (the default) does not work well for me because I use this a lot!
In normal mode, I use this to scroll the buffer (quite often).
In insert mode, I use to copy the text from above.
Having Emmet use this key mapping puts a delay on these actions.

<C-f> is one of the only key mappings that aren't used in insert mode
(according to :help index), but is used in normal mode to scroll
"forward" by one full page (aka pagedown).
Scolling by full pages tends to make me loose my orientation, so I tend
to use <C-d> instead of <C-f> which by default scrolls by half a page
instead.
Also, on many keyboards there is a "page down" button.

With all that in mind, as much as I dislike changing defaults, I think
its best to re-map this to <C-f>