Replace COC with neovim's builtin LSP
This also meant replacing coc-explorer with neo-tree as this was the only file
explorer that I could find that used floating windows!

This commit is also missing any autocomplete, however the built in omnicomplete
does work, although it seems to leave an empty 2 line window open after use.

I may need to also setup nvim-cmp (or similar) or figure out a way to stop that
omnicomplete window from hanging around.