Add icons to built in LSP and remove preview window
This is triggered by using <C-x><C-o> to open the omnicomplete menu (lsp
autocomplete), a preview window opens at the bottom but it's always blank in TS
projects, not sure if it works anywhere else.

Once you've selected the thing to auto complete, it doesn't go away so the
autocommand makes it disappear once the menu is closed.