Use LuaFormatter until Sumneko can take over
The formatter built into the sumneko lsp server would be ideal but it doesn't
seem to be able to format yet, maybe in a future release?

I tried stylua but it's very opinionated and does not allow much configuration,
and I didn't like the style.

LuaFormatter is much more configurable but I had a hard time getting the
luarocks install working and had to compile from source.

These little extra packages that are required by this new setup is going to be
annoying if I need to setup this stuff on a new machine and should probably be
scripted at some point.