Show parent-dir/ℹ️  for index.etc files
This is something I've wanted for a while!

When working in we codebases, it's common to call a file index.js.
When there are many tabs open with different index.js files open, it's hard to
differentiate unless you rename the tab, which I used to do with the Tabb.vim
plugin.

Most of the time, the parent directory name is good enough to describe what that
tab has open within it, even if it's multiple windows because normally the open
windows are the implementation, test, related styles, etc.

This change makes that the default, while still allowing the tab to be renamed
if that doesn't work.

I also used a nerd font glyph instead of the emoji at the top of this message
but thought it may not be readable on github, etc.