Don't automatically select auto complete items
This is pretty annoying when you've typed the text you want and go to insert a
new line by pressing enter but then it selects the first option in the
autocomplete list.

This would probably not be so bad if I didn't use enter as my selection key, but
enter just seems like the most intuitive key in this case.