Name snippets and prevent bug when not finishing
Before this change, if you started using a snippet but didn't finish jumping to
each insert node, but deleted that line instead, it would get very confused and
leave text on that line, and even when deleting this old text, it would come
back next time you interact with that line - very frustrating!

This change means it'll check if the snippet that was in progress has been
deleted every time the text changes or you leave insert mode.
Hopefully that doesn't slow things down but it seems much better now.