Install inc_rename.nvim and hook into Vimade
inc_rename will use LSP to perform renames but shows a preview as your typing,
like inccommand = "nosplit".

It uses vim.ui.input, which I have the dressing.nvim plugin providing a floating
window for.

Also, knowing it will live update, I was able to disable Vimade (the plugin that
fades the non-focused windows) when triggering the rename, then configured the
plugin to turn it back on after a rename has been performed.