Rollback toggling Vimade when inc_renaming
This would a nice addition but if a rename was started but then aborted, it
doesn't trigger the post_hook and never turns Vimade back on.

Might see if there is a more elegant way of doing in this in the future but it's
still pretty visible, even in faded windows.