Bringing back my custom tabby bar + nice colors
Having the line numbers column a different colour makes it easier to see
where the first column of the editable part starts.

This tab bar has a few advantages:
* Collapses `index.*` files to `<parent>/ℹ.<ext>` for legibility
* Shows how many splits are in a tab in a nicer way

While still providing the original features of:
* Filetype logo
* Modified indicator
* LSP diagnostics

The only drawback so far is that I've intentionally left out the close
button but also, the tabs don't seem to be clickable with the mouse
anymore.