Set blink-cmp to default keymaps
I got sick of the autocompletion appearing when I need to press enter
and it selecting a completion that I didn't want.

The standard "vim way" is to use <C-y> (for "yes") and while I don't
have muscle memory for this, it seems like a good default to get used
to.