Replace stow with chezmoi
Chezmoi (pronounced "shay-mwa") is one of the few dotfiles managers that
allows sourcing dotfiles from multiple sources in a layered fashion,
such that I can install these dotfiles from public github but also
override particular parts of it with a private, work-specific config.

It is very full featured but I'm only using the basics for now to keep
this config relatively portable incase I don't like it after a bit of
use.

Specifically leaving out the templating feature and it's way of running
scripts in favour of doing those bits manually for now.